backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-multipart
pytesseract
Pillow
google-generativeai
apscheduler
python-dotenv